Add svg assets for "Microsoft Account"
I'm not sure about the transparency here... it definitely violates the
simplicity of the rest of the icon set (not to mention Microsoft's brand
guidelines). But I couldn't think of any other way to make this actually
work in monochrome.

There is a "Windows" logo that works in monochrome (the window in
isometric perspective), but that is specific to Windows and Windows
Phone, not the Microsoft Account, which is what I needed it for.