Add Aurelia to json
Chose "ED2B88" color based on the buttons on the site; since the logo and most design elements have gradients it's tricky-feeling to choose one color. I think this choice is backed up though, because the "a" in the text logo is "ED2C88", just a tad brigher than the color I chose, and the "a" is the brightest letter in the logo. Let me know if you think otherwise.