Add Picarto.TV icon to json
The color is taken from the PNG icons on the press page of their site. In different contexts, they use different background colors for the logo.

It's specially called "Picarto.TV" on their site in multiple places- not "Picarto TV", "Picarto tv", "PicartoTV", or "PICARTO TV" (although some logos are in all-caps text). This can be verified as well on the press page of their website.