Add flattened version of the Sketch icon
Color is taken from the middle part of the bottom of the icon, as their
press page does not specify any colors to be used.

Note that their pess page does specify the following:

> We would appreciate it if you don’t modify or embellish Sketch brand
assets in any way, such as by changing the design or color of our
symbol, thanks.