Add GOG icon details
Added GOG icon details. The name is written as GOG as this is how it
appears as their website title and on their social media (even though
the logo is clearly lowercase). The (purple) color #86328A was picked
because GOG clearly has a purple esthetic, although the logos found at
the source all use (the gray) #333333.