Readd entrypoint for SVGO optimization to Dockerfile
Update the Dockerfile based on the original work in 
32993385daad3a66d6a2f8094c1dde5c33d2a03b by re-adding an ENTRYPOINT to 
the Dockerfile. This ENTRYPOINT makes it extremely easy to spin up a 
quick Docker container to optimize a single SVG (much simpler than my 
copy-in -> optimize -> copy-out approach).

The description for how to use the Docker image to run other NPM scripts 
has been updated accordingly. The provided command overrides the above 
ENTRYPOINT by simple starting a shell so the user can interact with the 
project.