Add a SVGO Docker image for Simpleicons formatting (#1532)
* Add a SVGO Docker image

* Update Dockerfile and .dockerignore

Update the Dockerfile to create a docker image that is generally 
applicable to run NPM commands, including but not limited to:

- npm run test
- npm run svgo
- npm run lint

Also updated the .dockerignore file to exclude:

- The node_modules folder
- Common Jekyll folders/files
- Files generated by the build script

The reason for choosing the alpine docker image (rather than a node 
docker image) is that the CLI out of the box is better.

* Add section on using Docker to Contributing Guidelines

* Readd entrypoint for SVGO optimization to Dockerfile

Update the Dockerfile based on the original work in 
32993385daad3a66d6a2f8094c1dde5c33d2a03b by re-adding an ENTRYPOINT to 
the Dockerfile. This ENTRYPOINT makes it extremely easy to spin up a 
quick Docker container to optimize a single SVG (much simpler than my 
copy-in -> optimize -> copy-out approach).

The description for how to use the Docker image to run other NPM scripts 
has been updated accordingly. The provided command overrides the above 
ENTRYPOINT by simple starting a shell so the user can interact with the 
project.

Co-authored-by: Eric Cornelisesn <ericornelissen@gmail.com>