Update build-package.js escape helper
The escape helper function in build-package.js is used to escape single
quotes in strings so that they don't cause trouble in JavaScript objects
that use single qoutes for keys and values. Though unlikely a key or
value containing the string "\'" (backslash-single quote) would break
the JS outputted by this script.

This potential bug was found by LGTM in [1].

---

1. https://lgtm.com/projects/g/simple-icons/simple-icons/snapshot/3580dd44ed2d48fc4dccb106165e7c5e871a0899/files/scripts/build-package.js?sort=name&dir=ASC&mode=heatmap#V31