Add GitHub Actions workflow for deployment
Add a GitHub Actions workflow file that does the deployment upon pushes
to master. Before actually deployin, the linters and test are ran, just
in case.

I did not test these exact deployments, but I based them on deployments
in other projects I have worked on.

Things not yet considered in this commit:
- Caching
- Git checkout depth