Switch from Travis CI to GitHub Actions (#4126)
* Add GitHub Actions workflow for verification

Add a GitHub Actions workflow file that does the verification of pushes
and pull requests. I.e., it runs the linters, runs the tests, and builds
the website. This workflow runs for all `pull_requests` and `push`es.

* Add GitHub Actions workflow for deployment

Add a GitHub Actions workflow file that does the deployment upon pushes
to master. Before actually deploying, the linters and test are ran, just
in case.

* Remove Travis CI configuration file

* Remove unnecessary quotes from existing workflows

* Add caching for "Build website" verification

* Update build badge in README

Co-authored-by: Álvaro Mondéjar <mondejar1994@gmail.com>