Use Personal Access token in release.yml (#4211)
Update release.yml, which is the workflow that creates release PRs and
merges them when approved, to use a Personal Access Token instead of
the standard GITHUB_TOKEN. I choose to name it RELEASE_TOKEN.

This change ensures that GitHub Actions workflows are triggered when
an action is performed by the Automated releases workflow. This is
needed because the publishing to NPM and the creation of a new GitHub
release should happen when there is a commit on `master`. If the release
PR is merged by the workflow using GITHUB_TOKEN, no GitHub Actions
workflows would be triggered.

Co-authored-by: Álvaro Mondéjar <mondejar1994@gmail.com>