Remove mention of GitPod
I believe the main advantage of GitPod was that it allows you to preview
a contribution in the context of the website. Now that the website is no
longer in the repository this is not a thing anymore (unless someone
can configure GitPod to pull in the website from its repo), so I removed
any mention of it.

I'm open to reverting this change if contributors still want to use
GitPod.