Update NodeJS usage documentation
First this updates the documentation to encourage users to get icons by
their slug rather than their title.

Second, I attempted to explain where the name comes from with a comment,
not sure if that works... I do think this could be valuable in general
(i.e. also for CDN usage and indiviual imports in NodeJS) but am unsure
how best to explain it...