Remove (old) website source files (#5553)
* Remove files directly related to the website

Remove all the files directly related to the Jekyll-based website in
this repository. This includes the HTML, CSS and JS for the website, the
assets that can be found in the images/ folder, Jekyll-related files
(_config.yml, Gemfile, Gemfile.lock) and the CNAME file.

* Remove website-related configurations

Remove website-related files from ignore files, update the EditorConfig
configuration to remove any mention of website files, and update
.gitattributes to omit removed files.

* Update issue templates

Remove the issue templates for issues regarding the website and replace
it with a issue templates configuration file [1] that provides a link to
a the equivalent issue template in the simple-icons-website repository.

--
1. https://docs.github.com/en/github/building-a-strong-community/configuring-issue-templates-for-your-repository


* Remove building website from verify workflow

* Remove mentions of website from Contributing Guidelines

* Remove mention of GitPod

I believe the main advantage of GitPod was that it allows you to preview
a contribution in the context of the website. Now that the website is no
longer in the repository this is not a thing anymore (unless someone
can configure GitPod to pull in the website from its repo), so I removed
any mention of it.

I'm open to reverting this change if contributors still want to use
GitPod.

* Add CI job for building the NodeJS package