Set up caching jest cache for tests in CI
Use actions/cache to cache the jest cache between runs. The cache will
be invalided if the package-lock.json file is updated OR after 7
days [1]. The cache directory for jest is now also configured to be in a
directory called ".jestcache" at the root of this project. This
directory will be ignored by git due to its addition to the .gitignore.

--
1. https://docs.github.com/en/actions/guides/caching-dependencies-to-speed-up-workflows#usage-limits-and-eviction-policy