Improve CI config: caching & refactoring (#5912)
* Set up caching jest cache for tests in CI

Use actions/cache to cache the jest cache between runs. The cache will
be invalided if the package-lock.json file is updated OR after 7
days [1]. The cache directory for jest is now also configured to be in a
directory called ".jestcache" at the root of this project. This
directory will be ignored by git due to its addition to the .gitignore.

--
1. https://docs.github.com/en/actions/guides/caching-dependencies-to-speed-up-workflows#usage-limits-and-eviction-policy

* Trigger jest cache in CI

* Include date in cache key

This way the cache is invalidated very day

* Add name for caching Jest's cache step

* Demonstrate test failure with cache enabled

* Revert ff915d466075d5292c7b5772406ff62fe1a3e220

* Cache dependencies

* Update Jest cache configuration

* Refactor and use Jest cache in Publish workfow

- Refactor out the duplicated sanity check in the publish workflow into
  a separate, depended upon, job.
- Use the Jest cache in the Publish workflow.

* Update linter job in CI to just run `npm run lint`

... whereas previously it ran each linter separately. This avoids having
to update the workflow file if a new linter is added (or a linter is
removed).

* Use run_number instead of date for jest cache in CI

* Simplify GitHub release publish job