Use PAT to avoid version bump commit problems due to branch protections
Following the instructions of stefanzweifel/git-auto-commit-action [1]
for pushing to branches with branch protections, this changes the create
release workflow to fetch the repo with a PAT (that must belong to an
admin for this to work). As a result, when pushing the version bump
commit, the required status checks are not required (provided the rules
are not applied to admins).

---
1. https://github.com/stefanzweifel/git-auto-commit-action#push-to-protected-branches